requirements.txt
flask
flask_cors
psycopg2-binary
mysql.connector
ibm_db
cx_Oracle
pyodbc
